Isolated all volumedriver calls
